# swagger.yaml
openapi: 3.0.3
info:
  title: Helpp API
  description: |
    API do projeto **Helpp** — plataforma de ajuda mútua que conecta pessoas que precisam de ajuda com quem pode ajudar.

    ## Autenticação
    A maioria dos endpoints protegidos usa **Bearer Token (JWT)**. O access token expira em **15 minutos**.
    Use `POST /auth/refresh` para renová-lo com o refresh token (validade de 30 dias).

    ## Sistema de Reciprocidade
    Cada usuário possui um saldo de reciprocidade:
    - `helpReceived` incrementa quando recebe uma ajuda concluída
    - `helpGiven` incrementa quando conclui uma ajuda como helper
    - Se `helpReceived > helpGiven`, o usuário está bloqueado de criar novos pedidos até ajudar alguém
  version: 1.0.0
  contact:
    name: Helpp
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        nickname:
          type: string
        cpfSuffix:
          type: string
          description: Últimos 3 dígitos do CPF
        phoneSuffix:
          type: string
          description: Últimos 4 dígitos do telefone
        city:
          type: string
        neighborhood:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        avatar:
          type: string
          description: Imagem em base64
        profileComplete:
          type: boolean
        helpReceived:
          type: integer
          description: Total de ajudas recebidas
          default: 0
        helpGiven:
          type: integer
          description: Total de ajudas dadas
          default: 0
...
    Message:
      type: object
      properties:
        _id:
          type: string
        matchId:
          type: string
        senderId:
          type: object
          properties:
            _id:
              type: string
            name:
              type: string
            nickname:
              type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
...
    VentQueueEntry:
      type: object
      properties:
        _id:
          type: string
        status:
          type: string
          enum: [waiting, matched]
        position:
          type: integer
          description: Posição na fila (1 = próximo a ser atendido)
          example: 3
...
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Busca perfil do usuário
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Perfil do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Usuário não encontrado

    patch:
      tags: [Users]
      summary: Atualiza perfil do usuário autenticado
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                cpfSuffix:
                  type: string
                phoneSuffix:
                  type: string
                city:
                  type: string
                neighborhood:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                avatar:
                  type: string
                  description: Imagem em base64
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Acesso negado
...
  /chat/{matchId}:
    get:
      tags: [Chat]
      summary: Busca mensagens de um match
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mensagens do match em ordem cronológica
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '403':
          description: Usuário não é participante do match
        '404':
          description: Match não encontrado

    post:
      tags: [Chat]
      summary: Envia mensagem no chat do match
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: matchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        '201':
          description: Mensagem enviada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Usuário não é participante do match
...
  /vent/queue:
    post:
      tags: [Vent (Desabafar)]
      summary: Entra na fila de espera (ou retorna posição se já estiver)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Entrada na fila criada ou existente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentQueueEntry'

    delete:
      tags: [Vent (Desabafar)]
      summary: Sai da fila de espera
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Saiu da fila com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Left queue
  /vent/queue/me:
    get:
      tags: [Vent (Desabafar)]
      summary: Consulta posição atual na fila
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Posição na fila
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentQueueEntry'
        '404':
          description: Usuário não está na fila